Change how Reddit access tokens get fetched
This greatly simplifies the setup of Reddit, by skipping the manual
creationg of access tokens and requiring only the client id and secret
to be added to the configuration.

An access token then gets automatically fetched by the function and
store in firebase for further usage. The token also gets renweed every
24 hours.

A downside here is that one cannot specify a scope of access, so the
script i.e. the access token basically has access to all possible API
functions.